+++ docs/serve.jl
@@ -1,4 +1,4 @@
-using CairoMakie
+import CairoMakie
 using Pollen
 using FastAI
 using FluxTraining