+++ src/datablock/checks.jl
@@ -8,18 +8,18 @@ function checkmethod_plot(
         sample = mocksample(method),
         devicefn = cpu,
         context = Training())
-    @testset "Plotting interface" begin
+    Test.@testset "Plotting interface" begin
         x, y = DLPipelines.encode(method, context, sample)
         ŷ = DLPipelines._predictx(method, model, x, devicefn)
 
-        @testset "plotsample!" begin
-            @test_nowarn plotsample(method, sample; resolution = (200, 200))
+        Test.@testset "plotsample!" begin
+            Test.@test_nowarn plotsample(method, sample; resolution = (200, 200))
         end
-        @testset "plotxy!" begin
-            @test_nowarn plotxy(method, x, y; resolution = (200, 200))
+        Test.@testset "plotxy!" begin
+            Test.@test_nowarn plotxy(method, x, y; resolution = (200, 200))
         end
-        @testset "plotprediction!" begin
-            @test_nowarn plotprediction(method, x, ŷ, y; resolution = (200, 200))
+        Test.@testset "plotprediction!" begin
+            Test.@test_nowarn plotprediction(method, x, ŷ, y; resolution = (200, 200))
         end
     end
 end