+++ src/datablock/wrappers.jl
@@ -3,6 +3,7 @@
 abstract type WrapperBlock <: AbstractBlock end
 
 wrapped(w::WrapperBlock) = w.block
+wrapped(b::Block) = b
 function setwrapped(w::WrapperBlock, b)
     return Setfield.@set w.block = b
 end
@@ -10,21 +11,98 @@ mockblock(w::WrapperBlock) = mockblock(wrapped(w))
 checkblock(w::WrapperBlock, data) = checkblock(wrapped(w), data)
 
 # If not overwritten, encodings are applied to the wrapped block
+"""
+    abstract type PropagateWrapper
 
-function encodedblock(enc::Encoding, wrapper::WrapperBlock)
+Defines the default propagation behavior of a `WrapperBlock` when
+an encoding is applied to it.
+
+Propagation refers to what happens when an encoding is applied to
+a `WrapperBlock`. If no `encode` method is defined for a wrapper block
+`wrapper`, `encode` is instead called on the wrapped block.
+Propagating the wrapper block means that the block resulting from
+encoding the wrapped block is rewrapped in `wrapper.`.
+
+```
+wrapper = Wrapper(block)
+# propagate
+encodedblock(enc, wrapper) = Wrapper(encodedblock(enc, wrapped(wrapper)))
+
+# don't propagate
+encodedblock(enc, wrapper) = encodedblock(enc, wrapped(wrapper))
+```
+
+The following wrapping behaviors exist:
+
+- `PropagateAlways`: Always propagate. This is the default behavior.
+- `PropagateNever`: Never propagate
+- `PropagateSameBlock`: Only propagate if the wrapped block is unchanged
+"""
+abstract type PropagateWrapper end
+
+struct PropagateAlways <: PropagateWrapper end
+struct PropagateSameBlock <: PropagateWrapper end
+struct PropagateNever <: PropagateWrapper end
+
+propagatewrapper(::WrapperBlock) = PropagateAlways()
+
+encodedblock(enc::Encoding, wrapper::WrapperBlock) =
+    encodedblock(enc, wrapper, propagatewrapper(wrapper))
+
+function encodedblock(enc::Encoding, wrapper::WrapperBlock, ::PropagateAlways)
     inner = encodedblock(enc, wrapped(wrapper))
     return isnothing(inner) ? nothing : setwrapped(wrapper, inner)
 end
-function decodedblock(enc::Encoding, wrapper::WrapperBlock)
+
+function encodedblock(enc::Encoding, wrapper::WrapperBlock, ::PropagateNever)
+    return encodedblock(enc, wrapped(wrapper))
+end
+
+function encodedblock(enc::Encoding, wrapper::WrapperBlock, ::PropagateSameBlock)
+    inner = encodedblock(enc, wrapped(wrapper))
+    inner == wrapped(block) && return setwrapped(wrapper, inner)
+    return inner
+end
+
+decodedblock(enc::Encoding, wrapper::WrapperBlock) =
+    decodedblock(enc, wrapper, propagatewrapper(wrapper))
+
+function decodedblock(enc::Encoding, wrapper::WrapperBlock, ::PropagateAlways)
     inner = decodedblock(enc, wrapped(wrapper))
     return isnothing(inner) ? nothing : setwrapped(wrapper, inner)
 end
+
+function decodedblock(enc::Encoding, wrapper::WrapperBlock, ::PropagateNever)
+    return decodedblock(enc, wrapped(wrapper))
+end
+
+function decodedblock(enc::Encoding, wrapper::WrapperBlock, ::PropagateSameBlock)
+    inner = decodedblock(enc, wrapped(wrapper))
+    inner == wrapped(block) && return setwrapped(wrapper, inner)
+    return inner
+end
+
+
+# Encoding and decoding, if not overwritten for specific wrapper, are fowarded
+# to wrapped block.
+
 function encode(enc::Encoding, ctx, wrapper::WrapperBlock, data; kwargs...)
     return encode(enc, ctx, wrapped(wrapper), data; kwargs...)
 end
+
 function decode(enc::Encoding, ctx, wrapper::WrapperBlock, data; kwargs...)
     return decode(enc, ctx, wrapped(wrapper), data; kwargs...)
 end
+
+
+# Training interface
+
+blockbackbone(wrapper::WrapperBlock) = blockbackbone(wrapped(wrapper))
+blockmodel(wrapper::WrapperBlock, out, args...) = blockmodel(wrapped(wrapper), out, args...)
+blockmodel(in::Block, out::WrapperBlock, args...) = blockmodel(in, wrapped(out), args...)
+
+blocklossfn(wrapper::WrapperBlock, out) = blocklossfn(wrapped(wrapper), out)
+blocklossfn(in::Block, out::WrapperBlock) = blocklossfn(in, wrapped(out))
 
 # ## Named
 
@@ -34,10 +112,10 @@ end
 Wrapper `Block` to attach a name to a block. Can be used in conjunction
 with [`Only`](#) to apply encodings to specific blocks only.
 """
-struct Named{Name, B<:AbstractBlock} <: WrapperBlock
+struct Named{Name,B <: AbstractBlock} <: WrapperBlock
     block::B
 end
-Named(name::Symbol, block::B) where {B<:AbstractBlock} = Named{name, B}(block)
+Named(name::Symbol, block::B) where {B <: AbstractBlock} = Named{name,B}(block)
 
 
 # the name is preserved through encodings and decodings
@@ -63,7 +141,7 @@ samples. The blocks `(Image{2}(), BoundingBox{2}()` imply that there is exactly
 one bounding box for every image, which is not the case. Instead you
 would want to use `(Image{2}(), Many(BoundingBox{2}())`.
 """
-struct Many{B<:AbstractBlock} <: WrapperBlock
+struct Many{B <: AbstractBlock} <: WrapperBlock
     block::B
 end
 
@@ -79,7 +157,7 @@ end
 function FastAI.decode(enc::Encoding, ctx, many::Many, datas)
     return map(datas) do data
         decode(enc, ctx, wrapped(many), data)
-    end
+end
 end
 
 
@@ -87,17 +165,16 @@ end
 
 """
     Only(name, encoding)
-    Only(block, encoding)
 
 Wrapper that conditionally applies `encoding` only if the block
-equals `block` or is a `Named{name}`.
+is a `Named{name}`.
 """
-struct Only{Name, E<:Encoding} <: StatefulEncoding
+struct Only{Name,E <: Encoding} <: StatefulEncoding
     encoding::E
 end
 
 function Only(name::Symbol, encoding::E) where E
-    return Only{name, E}(encoding)
+    return Only{name,E}(encoding)
 end
 
 
@@ -114,7 +191,7 @@ function encode(
         block::Named{Name},
         data;
         state=encodestate(only, context, block, data)) where Name
-    return encode(only.encoding, context, block, data; state = state)
+    return encode(only.encoding, context, block, data; state=state)
 end
 
 
@@ -124,5 +201,5 @@ function decode(
         block::Named{Name},
         data;
         state=decodestate(only, context, block, data)) where Name
-    return decode(only.encoding, context, block, data; state = state)
+    return decode(only.encoding, context, block, data; state=state)
 end