+++ test/methods/imagesegmentation.jl
@@ -8,7 +8,14 @@ include("../imports.jl")
         DLPipelines.checkmethod_core(method)
         @test_nowarn methodlossfn(method)
         @test_nowarn methodmodel(method, Models.xresnet18())
-        FastAI.checkmethod_plot(method)
+        @testset "Show backends" begin
+            @testset "ShowText" begin
+                FastAI.test_method_show(method, ShowText(Base.DevNull()))
+            end
+            @testset "ShowMakie" begin
+                FastAI.test_method_show(method, ShowMakie())
+            end
+        end
     end
     @testset ExtendedTestSet "3D" begin
         method = BlockMethod(